app/server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
numpy==2.2.4
open-clip-torch==2.32.0
pillow==11.1.0
qdrant-client==1.14.2
torch==2.6.0
